bumped version

+++ db_mutex/db_mutex.py
@@ -1,10 +1,14 @@
 from datetime import datetime, timedelta
 import functools
+import logging
 
 from django.conf import settings
 from django.db import transaction, IntegrityError
 
 from .models import DBMutex
+
+
+LOG = logging.getLogger(__name__)
 
 
 class DBMutexError(Exception):
@@ -63,9 +67,10 @@ class db_mutex(object):
     """
     mutex_ttl_seconds_settings_key = 'DB_MUTEX_TTL_SECONDS'
 
-    def __init__(self, lock_id):
+    def __init__(self, lock_id, suppress_acquisition_exceptions=False):
         self.lock_id = lock_id
         self.lock = None
+        self.suppress_acquisition_exceptions = suppress_acquisition_exceptions
 
     def get_mutex_ttl_seconds(self):
         """
@@ -102,16 +107,20 @@ class db_mutex(object):
             with transaction.atomic():
                 self.lock = DBMutex.objects.create(lock_id=self.lock_id)
         except IntegrityError:
-            raise DBMutexError('Could not acquire lock: {0}'.format(self.lock_id))
+            if self.suppress_acquisition_exceptions:
+                LOG.error('Could not acquire lock: {0}'.format(self.lock_id))
+            else:
+                raise DBMutexError('Could not acquire lock: {0}'.format(self.lock_id))
 
     def stop(self):
         """
         Releases the db mutex lock. Throws an error if the lock was released before the function finished.
         """
-        if not DBMutex.objects.filter(id=self.lock.id).exists():
-            raise DBMutexTimeoutError('Lock {0} expired before function completed'.format(self.lock_id))
-        else:
-            self.lock.delete()
+        if self.lock:
+            if not DBMutex.objects.filter(id=self.lock.id).exists():
+                raise DBMutexTimeoutError('Lock {0} expired before function completed'.format(self.lock_id))
+            else:
+                self.lock.delete()
 
     def decorate_callable(self, func):
         """
